deco_irq: y doesn't fit in a byte (nw)

+++ src/mame/machine/deco_irq.cpp
@@ -90,7 +90,7 @@ void deco_irq_device::device_reset()
 TIMER_CALLBACK_MEMBER( deco_irq_device::scanline_callback )
 {
 	const rectangle visible = m_screen->visible_area();
-	uint8_t y = m_screen->vpos();
+	uint16_t y = m_screen->vpos();
 
 	// raster irq?
 	if (m_raster_irq_scanline > 0 && m_raster_irq_scanline < 240 && y == (m_raster_irq_scanline - 1))
